gh #186: dsAudio L3 developement code
- Folder restructure
