+++ host/tests/L3_TestCases/dsAudio/dsAudio_L3_testSetup.yml
@@ -1,0 +1,120 @@
+dsAudio:
+  description: "dsAudio Device Settings test setup"
+  assets:
+    device:
+      defaults: &defaults
+        artifacts:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/bin/hal_test"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/bin/libut_control.so"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/bin/Sink_AudioSettings.yaml"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/bin/run.sh"
+        execute:
+          - "chmod +x /opt/HAL_L3/hal_test"
+          - "chmod +x /opt/HAL_L3/run.sh"
+        streams:
+      test01_EnableDisableAndVerifyAudioPortStatus:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test02_PortConnectionStatus:
+        <<: *defaults
+        streams:
+      test03_MS12AudioCompression:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test04_MS12DialogueEnhancer:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test05_MS12DolbyVolume:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test06_MS12IntelligentEqualizer:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test07_MS12Volumeleveller:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test08_MS12BassEnhancer:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test09_MS12SurroundDecoder:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test10_MS12DRCMode:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test11_MS12SurroundVirtualizer:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test12_MS12MISteering:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test13_MS12GraphicEqualizer:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test14_MS12LEConfig:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test15_ARCPort:
+        <<: *defaults
+        streams: []
+      test16_ARCSAD:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test17_OutputMode:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.wav"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.eac3"
+      test18_AudioLevel:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test19_AudioGain:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test20_MuteUnMute:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test21_AudioDelay:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/AVsync_aac_200_48k_ATSC_m2v_29fps.trp"
+      test22_AudioFormat:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.wav"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test23_AssociateMix:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test24_PrimarySecondaryLanguage:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
+      test25_AudioMix:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/primary_audio_48k_2ch.ac3"
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/secondary_audio_48k_2ch.wav"
+      test26_MS12AudioProfiles:
+        <<: *defaults
+        streams:
+          - "http://hydra.bskyb.com/shared/Projects/RDK_HAL/HAL_Testing/Element/Demo/DeviceSettings/streams/tones_string_48k_stereo.ac3"
