--- host/tests/dsAudio_L3_Tests/dsAudio_L3_testSetup.yml
@@ -1,120 +1,0 @@
-dsAudio:
-  description: "dsAudio Device Settings test setup"
-  assets:
-    device:
-      defaults: &defaults
-        artifacts:
-          - "<PATH>/bin/hal_test"
-          - "<PATH>/bin/libut_control.so"
-          - "<PATH>/bin/Sink_AudioSettings.yaml"
-          - "<PATH>/bin/run.sh"
-        execute:
-          - "chmod +x <PATH>/hal_test"
-          - "chmod +x <PATH>/run.sh"
-        streams:
-      test01_EnableDisableAndVerifyAudioPortStatus:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test02_PortConnectionStatus:
-        <<: *defaults
-        streams:
-      test03_MS12AudioCompression:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test04_MS12DialogueEnhancer:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test05_MS12DolbyVolume:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test06_MS12IntelligentEqualizer:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test07_MS12Volumeleveller:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test08_MS12BassEnhancer:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test09_MS12SurroundDecoder:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test10_MS12DRCMode:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test11_MS12SurroundVirtualizer:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test12_MS12MISteering:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test13_MS12GraphicEqualizer:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test14_MS12LEConfig:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test15_ARCPort:
-        <<: *defaults
-        streams: []
-      test16_ARCSAD:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test17_OutputMode:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.wav"
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-          - "<PATH>/streams/tones_string_48k_stereo.eac3"
-      test18_AudioLevel:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test19_AudioGain:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test20_MuteUnMute:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test21_AudioDelay:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/AVsync_aac_200_48k_ATSC_m2v_29fps.trp"
-      test22_AudioFormat:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.wav"
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-
-      test23_AssociateMix:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test24_PrimarySecondaryLanguage:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test25_AudioMix:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
-      test26_MS12AudioProfiles:
-        <<: *defaults
-        streams:
-          - "<PATH>/streams/tones_string_48k_stereo.ac3"
